+++ build-binary-from-source.sh
@@ -1,4 +1,4 @@
 cd ..
-R CMD INSTALL --build --no-multiarch ooacquire_0.3.2-1.tar.gz
+R CMD INSTALL --build --no-multiarch ooacquire_0.3.3.tar.gz
 cd ./ooacquire